UBUNTU: SAUCE: i2c: designware: Add disable runtime pm quirk
BugLink: https://bugs.launchpad.net/bugs/1833484

Dell machines come with goodix touchpad IC suffer from the double click
issue if the Designware I2C adapter enters runtime suspend.

It's because the goodix re-assert the interrupt if host doesn't read the
data within 100ms and desiginware takes a longer time to wake up from
runtime suspend. In the case, it got a second interrupt during
resuming, so it thinks it's a double click.

There is no simple way to fix this, it's a firmware issue and goodix
agrees to fix this in their firmware on next release, but this issue is
still affects the machines that don't come with an updated firmware. So,
add a quirk to mark those machines and avoid the designware to enter
runtime suspend.

Signed-off-by: AceLan Kao <acelan.kao@canonical.com>
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>